+++ app/preloaders/api/v2/contacts/people_preloader.rb
@@ -1,4 +1,7 @@
 class Api::V2::Contacts::PeoplePreloader < ApplicationPreloader
   ASSOCIATION_PRELOADER_MAPPING = { family_relationships: Api::V2::Contacts::People::RelationshipsPreloader }.freeze
-  FIELD_ASSOCIATION_MAPPING = { avatar: [:primary_picture, :facebook_account] }.freeze
+  FIELD_ASSOCIATION_MAPPING = {
+    avatar: [:primary_picture, :facebook_account],
+    parent_contacts: :contacts
+  }.freeze
 end